Refactor authentication and authorization

+++ src/controller/OrderController.ts
@@ -2,6 +2,8 @@ import mariadb from "mysql2/promise";
 import { Request, Response } from "express";
 import { config } from "dotenv";
 import { StatusCodes } from "http-status-codes";
+import ensureAuthorization from "src/auth";
+import jwt from "jsonwebtoken";
 
 config();
 const order = async (req: Request, res: Response): Promise<void> => {
@@ -12,44 +14,57 @@ const order = async (req: Request, res: Response): Promise<void> => {
     database: process.env.DB_Database,
     dateStrings: true,
   });
+  let authorization = ensureAuthorization(req, res);
 
-  const { items, delivery, totalQuantity, totalPrice, userId, firstBookTitle } =
-    req.body;
+  if (authorization instanceof jwt.TokenExpiredError) {
+    res.status(StatusCodes.UNAUTHORIZED).json({ message: "Token expired" });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    res.status(StatusCodes.BAD_REQUEST).json({ message: "잘못된 토큰입니다." });
+  } else {
+    const { items, delivery, totalQuantity, totalPrice, firstBookTitle } =
+      req.body;
 
-  // delivery 테이블 삽입
-  let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
-  let values = [delivery.address, delivery.receiver, delivery.contact];
-  let [results] = await conn.execute(sql, values);
-  //@ts-ignore
-  let delivery_id = results.insertId;
+    // delivery 테이블 삽입
+    let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
+    let values = [delivery.address, delivery.receiver, delivery.contact];
+    let [results] = await conn.execute(sql, values);
+    //@ts-ignore
+    let delivery_id = results.insertId;
 
-  // orders 테이블 삽입
-  sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
-          VALUES (?,?,?,?,?)`;
-  values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-  [results] = await conn.execute(sql, values);
-  //@ts-ignore
-  let order_id = results.insertId;
+    // orders 테이블 삽입
+    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id)
+        VALUES (?,?,?,?,?)`;
+    values = [
+      firstBookTitle,
+      totalQuantity,
+      totalPrice,
+      authorization.id,
+      delivery_id,
+    ];
+    [results] = await conn.execute(sql, values);
+    //@ts-ignore
+    let order_id = results.insertId;
 
-  //items를 가지고, 장바구니에서 book_id, quantity 조회
-  sql = `SELECT book_id, quantity FROM cartitems WHERE id IN (?)`;
-  let [orderItems, fields] = await conn.query(sql, [items]);
+    //items를 가지고, 장바구니에서 book_id, quantity 조회
+    sql = `SELECT book_id, quantity FROM cartitems WHERE id IN (?)`;
+    let [orderItems, fields] = await conn.query(sql, [items]);
 
-  // orderedBook 테이블 삽입
-  sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
+    // orderedBook 테이블 삽입
+    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
 
-  values = [];
-  if (orderItems instanceof Array) {
-    orderItems.forEach((item: any) => {
-      values.push([order_id, item.book_id, item.quantity]);
-    });
+    values = [];
+    if (orderItems instanceof Array) {
+      orderItems.forEach((item: any) => {
+        values.push([order_id, item.book_id, item.quantity]);
+      });
+    }
+
+    results[0] = await conn.query(sql, [values]);
+
+    let result = await deleteCartItems(conn, items);
+
+    res.status(StatusCodes.OK).json(result);
   }
-
-  results[0] = await conn.query(sql, [values]);
-
-  let result = await deleteCartItems(conn, items);
-
-  res.status(StatusCodes.OK).json(result);
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -77,7 +92,6 @@ const getOrders = async (req: Request, res: Response): Promise<void> => {
 };
 
 const getOrderDetail = async (req: Request, res: Response): Promise<void> => {
-  const { id } = req.params;
   const conn = await mariadb.createConnection({
     host: process.env.DB_Host,
     user: process.env.DB_User,
@@ -90,7 +104,7 @@ const getOrderDetail = async (req: Request, res: Response): Promise<void> => {
               FROM orderedBook LEFT JOIN books
               ON orderedBook.book_id = books.id
               WHERE order_id = ?`;
-  let [rows, fields] = await conn.query(sql, [id]);
+  let [rows, fields] = await conn.query(sql, [req.params.id]);
   res.status(StatusCodes.OK).json(rows);
 };
